split into sub-components
