+++ garden/src/app/router.tsx
@@ -8,7 +8,7 @@ import RootLayout from "@/components/Layout";
 
 import { useGlobusAuth } from "@globus/react-auth-context";
 import SuperuserRoute from "@/components/SuperuserRoute";
-import { UnifiedManagmentInterface } from "@/features/ugmi/UnifiedManagementInterface";
+import { UnifiedManagementInterface } from "@/features/ugmi/UnifiedManagementInterface";
 
 const EntrypointTombstonePage = lazy(() => import("@/components/EntrypointTombstonePage"));
 const CreateGardenPage = lazy(() => import("@/features/gardens/components/create/CreateGardenPage"));
@@ -84,7 +84,7 @@ const Router: React.FC = () => {
         <Route element={<SuperuserRoute />}>
           <Route path="metrics" element={<WrappedLazyComponent child={<MetricsDashboard />} />}></Route>
         </Route>
-        <Route path="ugmi" element={<UnifiedManagmentInterface />} />
+        <Route path="ugmi" element={<UnifiedManagementInterface />} />
       </Route>
       <Route path="*" element={<NotFoundPage />} />
     </Routes >
