+++ garden/src/features/ugmi/components/index.ts
@@ -1,0 +1,22 @@
+// Panel Components
+export { LeftSidePanel } from "./LeftSidePanel";
+export { RightSidePanel } from "./RightSidePanel";
+export { MainContentPanel } from "./MainContentPanel";
+
+// Garden Components
+export { SavedGardensPanel } from "./SavedGardensPanel";
+export { MyGardensPanel } from "./MyGardensPanel";
+export { PublishedGardensPanel } from "./PublishedGardensPanel";
+
+// Function Library Components
+export { MyFunctionLibraryView } from "./MyFunctionLibraryView";
+export { DeploymentGroup } from "./DeploymentGroup";
+export { FunctionItem } from "./FunctionItem";
+
+// Metadata and User Components
+export { MetadataPanel } from "./MetadataPanel";
+export { UserInfoPanel } from "./UserInfoPanel";
+
+// Content Components
+export { UnifiedGardenContent } from "./UnifiedGardenContent";
+export { UnifiedFunctionContent } from "./UnifiedFunctionContent";
